Remove the heal delay.
Using a different delay for each type of heal auto status leads to scenarios where RSR makes
irrational decisions. For example, in a situation where both
canHealSingleSpell and canHealAreaSpell become true at the same time
(like a hard-hitting AOE), it may decide to do a single target heal just
because that delay happened to be shorter.