bypass ttk check
A lot of GCDs couldn't be used at the end of the fights due to TTK check. This is a bypass, so no need to change every job's rotation file in RSR. Turn on true when used in RR if desired, otherwise it doesn't affect existing code because it's false by default.