Rename and update NPC Heal/Raise config setting
Renamed `_friendlyPartyNPCHealRaise` to `_friendlyPartyNPCHealRaise2` and changed its default value from `true` to `false`. Updated the description to indicate that it is experimental and may cause issues in some fights. Updated all references in `DataCenter.cs`, `ObjectHelper.cs`, and `RotationConfigWindow.cs` to use the new setting name. Modified the UI text and button functionality in `RotationConfigWindow.cs` to reflect the new setting.