PCT low level hammer fix
At first glance `burstTimingCheckerStriking` should be irrelevant when player's level isn't high enough to use `Starry Muse`, so i made it be true if that happens.
This fixes the hammer usage for levels <70 and the rotation doesn't seem to be negatively affected when level is high enough for `Starry Muse` to be unlocked.

Tested on Lv 60,70 SSS. Similar results for changed and current rotation.