Update NPC Heal/Raise config and add new status entries
Renamed `_friendlyPartyNPCHealRaise2` to `_friendlyPartyNPCHealRaise3` in `Configs` class and updated references in `ObjectHelper`, `RotationConfigWindow`, and `TargetUpdater` to reset default config for all users to true. Refactored `IsCastingAreaVfx` in `DataCenter` to include additional checks. Reformatted XML comments in `StatusGetter` and added new status entries to `Resources.resx` with summaries and links to garlandtools.org to allow for tracking of all statuses.